pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geocaptcha-shillelagh"
version = "0.1.0"
description = "Shillelagh adapter for the GeoCaptcha REST API (IGNF)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["shillelagh", "geocaptcha", "superset", "sql", "adapter", "ignf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "shillelagh>=1.4",
    "python-dateutil>=2.8",
    "requests-cache>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-mock>=3",
    "responses>=0.25",
]

[project.entry-points."shillelagh.adapter"]
geocaptchasessionadapter = "geocaptcha_shillelagh.session:GeocaptchaSessionAdapter"
geocaptchacuseradapter = "geocaptcha_shillelagh.cuser:GeocaptchaCuserAdapter"
geocaptchakingpinadapter = "geocaptcha_shillelagh.kingpin:GeocaptchaKingpinAdapter"


[project.urls]
Homepage = "https://github.com/ledav-perso/geocaptcha-shillelagh"
Repository = "https://github.com/ledav-perso/geocaptcha-shillelagh"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
